+++ packages/terra-scroll/CHANGELOG.md
@@ -5,7 +5,7 @@
 * Added
   * Added styles when the scroll is focussed
 * Changed
-  * Moved the focus style changes from terra-scroll to terra-content-continer
+  * Removed the focus style changes added for terra-scroll.
 
 ## 2.34.0 - (February 15, 2023)